MH-271: move moderated fields to moderation module and clean up old code
